moving styling to js file
