+++ client/src/components/Navbar.js
@@ -1,25 +1,135 @@
+/**
+ * Navbar Component
+ *
+ * Component that renders the navbar
+ *
+ * Pages will be a dictionary with {key: value} = {Page Name: Path to page}
+ * The pages prop will contain the elements of the navbar and their routes
+ *
+ * Example usage:
+ *   <Navbar
+ *     pages={{
+ *       "Pending Applications": "/dashboard",
+ *       "Current Fosters": "/fosters",
+ *       "Calendar": "/calendar",
+ *     }}
+ *   />
+ */
+
 import React, { useState, useEffect } from "react";
-import { Nav, NavLink, NavMenu, ToggleNavbar, Burger, SignOut } from "./Navbar_styled";
-import burger from "./burger.png";
+import { NavLink as Link } from "react-router-dom";
+import styled, { keyframes } from "styled-components";
+import burger from "../images/burger.png";
+
+// Styled components
+const slideIn = keyframes`
+ 0% { right: -50%; }
+ 20% { right: -40%; }
+ 40% { right: -30%; }
+ 60% { right: -20%; }
+ 80% { right: -10%; }
+ 100% { right: 0%; }
+`;
+
+export const Nav = styled.nav`
+  position: absolute;
+  width: max(250px, 16vw);
+  height: 85%;
+  top: 15%;
+  background-color: #000000;
+
+  @media screen and (max-width: 750px) {
+    top: 15%;
+    right: 0%;
+    width: min(300px, 45vw);
+    z-index: 5;
+
+    animation-name: ${slideIn};
+    animation-duration: 0.5s;
+    animation-iteration-count: 1;
+    }
+  }
+`;
+
+export const ToggleNavbar = styled.button`
+  display: none;
+
+
+  @media screen and (max-width: 750px) {
+    display: inline;
+    position: absolute;
+    right: 5%;
+    top: 5%;
+
+    background: none;
+    color: inherit;
+    border: none;
+    padding: 0;
+    font: inherit;
+    cursor: pointer;
+    outline: inherit;
+
+}
+  }
+`;
+
+export const Burger = styled.img``;
+
+export const NavMenu = styled.div`
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+  align-items: center;
+`;
+
+export const NavLink = styled(Link)`
+  text-decoration: none;
+
+  font-family: Myriad Pro;
+  font-style: normal;
+  font-weight: bold;
+  font-size: max(18px, min(24px, 2vw));
+  line-height: 29px;
+  text-align: center;
+  color: #ffffff;
+
+  width: 100%;
+  padding: 25px 0px;
+
+  &.active {
+    background-color: #8dc442;
+  }
+  @media screen and (max-width: 750px) {
+    font-size: min(18px, 4vw);
+  }
+`;
+
+export const SignOut = styled.button`
+  display: none;
+  @media screen and (max-width: 750px) {
+    display: block;
+    text-decoration: none;
+
+    font-family: Myriad Pro;
+    font-style: normal;
+    font-weight: bold;
+    font-size: max(14px, min(18px, 4vw));
+    line-height: 29px;
+    background-color: #000000;
+    background: none;
+    color: #ffffff;
+    border: none;
+    padding: 0;
+    cursor: pointer;
+
+    position: absolute;
+    bottom: 0;
+
+    padding: 25px 0px;
+  }
+`;
 
 function Navbar(props) {
-  // Pages will be a dictionary with {key: value} = {Page Name: Path to page}
-  /*
-  Example usage:
-    At top of file:
-    import Navbar from "../components/Navbar";
-
-    In JSX:
-    <Navbar
-      pages={{
-        "Pending Applications": "/dashboard",
-        "Current Fosters": "/fosters",
-        "Calendar": "/calendar",
-      }}
-    />
-    
-    The pages prop will contain the elements of the navbar and their routes
-  */
   const [renderNav, setRenderNav] = useState(false);
   const [screenWidth, setScreenWidth] = useState(window.innerWidth);
 
@@ -42,7 +152,7 @@ function Navbar(props) {
   return (
     <>
       <ToggleNavbar onClick={() => setRenderNav(!renderNav)}>
-        <Burger src={burger} />
+        <Burger src={burger} alt="burger" />
       </ToggleNavbar>
 
       {/* Conditionally Rendered Navigation Panel */}
@@ -56,14 +166,10 @@ function Navbar(props) {
             ))}
             {screenWidth < 750 || renderNav ? (
               <SignOut onClick={() => logout()}>Sign Out</SignOut>
-            ) : (
-              <></>
-            )}
+            ) : undefined}
           </NavMenu>
         </Nav>
-      ) : (
-        <></>
-      )}
+      ) : undefined}
     </>
   );
 }
